# SYCL: Unroll shuffle loops for top-level parallel_reduce and parallel_scan