--- a/core/src/SYCL/Kokkos_SYCL_ParallelScan_Range.hpp
+++ b/core/src/SYCL/Kokkos_SYCL_ParallelScan_Range.hpp
@@ -77,7 +77,7 @@ void workgroup_scan(sycl::nd_item<dim> item, const FunctorType& final_reducer,
             std::min(local_range, n_active_subgroups - round * local_range);
         auto local_sg_value = local_mem[idx < n_active_subgroups ? idx : 0];
 #if defined(KOKKOS_ARCH_INTEL_GPU) || defined(KOKKOS_IMPL_ARCH_NVIDIA_GPU)
-        auto shuffle_combine = [&](int stride) {
+        auto shuffle_combine_sg = [&](int stride) {
           if (stride < upper_bound) {
             auto tmp = sg.shuffle_up(local_sg_value, stride);
             if (id_in_sg >= stride) {
@@ -88,12 +88,12 @@ void workgroup_scan(sycl::nd_item<dim> item, const FunctorType& final_reducer,
             }
           }
         };
-        shuffle_combine(1);
-        shuffle_combine(2);
-        shuffle_combine(4);
-        shuffle_combine(8);
-        shuffle_combine(16);
-        shuffle_combine(32);
+        shuffle_combine_sg(1);
+        shuffle_combine_sg(2);
+        shuffle_combine_sg(4);
+        shuffle_combine_sg(8);
+        shuffle_combine_sg(16);
+        shuffle_combine_sg(32);
 #else
         for (int stride = 1; stride < upper_bound; stride <<= 1) {
           auto tmp = sg.shuffle_up(local_sg_value, stride);