Similar to #6562, unrolling the loops for shuffles improves performance significantly. For a simple dot product benchmark we are seeing on Intel PVC

| elements  | old shuffle (GB/s) | new shuffle (GB/s) | local memory (GB/s) |
| -- | -- | -- | -- |
| 2 |    0.00195408 | 0.00196642 | 0.00195283
| 4 |    0.00416202 | 0.00425014 | 0.00421528
| 8 |    0.00820138 | 0.00832427 | 0.0083891
| 16 |     0.0138029 | 0.0146039 | 0.0142298
| 32 |    0.0294672 | 0.0335499 | 0.0324571
| 64 |    0.0581855 | 0.064666 | 0.0623012
| 128 |  0.119849 | 0.133714 | 0.125332
| 256 |     0.229995 | 0.258772 | 0.237334
| 512 |    0.465701 | 0.509693 | 0.440304
| 1024 |  0.855585 | 1.01661 | 0.734111
| 2048 |    1.49275 | 1.72464 | 1.47962
| 4096 |    2.88409 | 3.56523 | 3.15379
| 8192 |    5.46391 | 6.73428 | 6.30253
| 16384 |    10.6889 | 13.2914 | 12.1061
| 32768 |   20.9248 | 26.2938 | 24.5314
| 65536 |   40.7319 | 49.5028 | 46.1789
| 131072 |   70.2748 | 95.7102 | 86.0199
| 262144 |     103.283 | 143.788 | 130.812
| 524288 |   160.729 | 289.661 | 259.618
| 1048576 |    219.284 | 402.79 | 349.205
| 2097152 |    420.953 | 750.37 | 656.928
| 4194304 |    776.545 | 1291.27 | 1196.56
| 8388608 |     1352.93 | 2073.44 | 1999.47
| 16777216 |   2125.82 | 2919.15 | 2990.47
| 33554432 |    1646.12 | 1686.77 | 1599.83
| 67108864 |    1621.1 | 1643.43 | 1587.29
| 134217728 |      1613.23 | 1607.5 | 1580.22
| 268435456 |    1533.8 | 1537.64 | 1532.46
| 536870912 |    1638.35 | 1641.31 | 1637.63
| 1073741824 |   1671.84 | 1670.2 | 1676.66

meaning that with this change we get similar performance between the local memory variant (needed for array reductions) and the shuffle variant. Thus, this allows for reviving the previously dead code path.
Note that `parallel_scan` only has a shuffle variant (and doesn't support array scans).