@ndellingwood 2024-05-01
changelog entry needed for this update to the 4.4 tracker https://github.com/kokkos/kokkos/issues/6914